refactor(dts): make `HighlightingAssets` cacheable (#8)
This is a relatively big refactor which replaces the usage of
`bat::PrettyPrinter` with a wrapper around `bat::controller::Controller`
to have better control of loading the highlighting assets as this is a
very heavy operations and we only ever want to do it once during the
whole runtime of `dts`.

This will enable some performance improvements later. Along the way I
added some defaulting here and there and made theme selection
case-insensitive for convenience.