feat: add custom `Value` impl behind feature switch
This is meant as a replacement for the internal usage of
`serde_json::Value`. The main issue with `serde_json::Value` is that we
are now at a point where we need features that it does not have and
which we cannot implement cleanly because `serde_json` is a foreign
crate.

For example, `serde_json::Value` is neither `Hash` nor `PartialOrd` or
`Ord`. We cannot implement these traits for it because both the type and
the traits are foreign.

For the latter two we did a nasty workaround. Other features like
deep merging we added via a `ValueExt` trait.