feat(dts-core): add `Transform` trait (#14)
The `Transformation` type got quite large and is not flexible enough for
some transformations that will be added in the future.

To fix this, the `Transform` trait was added and dynamic dispatch over
this type will be used. This will allow for a more modular design where
specialized transformations can have their dedicated type.