feat(dts-core/transform): nested expressions and `mutate` (#15)
This PR adds `mutate` and the code to nest expressions. 

Example:

```
mutate("$.users[*].friends", sort())
```

Only sorts each user's friends by name but not the rest of the data structure.

This enables a whole lot of new ways to compose transformation functions and enables new use cases, for example:

```
replace("$.users[*].friends", ["paul", "dave"])
```

This one does not work right now because neither `replace` nor array syntax is supported, but it'll be fairly easy to implement after this PR is merged.